refactor: monitoring value use case and hooks, get dataset name from config
